feat: 初始化脚本默认不重置venv环境，仅在build调用时重置

+++ _init_venv_and_requirements.py
@@ -9,7 +9,7 @@ from util import bypass_proxy, show_head_line
 
 
 def init_venv_and_requirements(
-    venv_path=".venv", requirements_path="requirements.txt", disable_douban=False, enable_proxy=False
+    venv_path=".venv", requirements_path="requirements.txt", disable_douban=False, enable_proxy=False, reset_venv=False
 ):
     if not enable_proxy:
         logger.info("当前已无视系统代理")
@@ -20,8 +20,9 @@ def init_venv_and_requirements(
     py_path = os.path.join(pyscript_path, "python")
     pip_path = os.path.join(pyscript_path, "pip")
 
-    show_head_line("先清空原来的环境，确保每次从头开始准备环境，避免莫名其妙的问题", color("bold_yellow"))
-    shutil.rmtree(venv_path, ignore_errors=True)
+    if reset_venv:
+        show_head_line("先清空原来的环境，确保每次从头开始准备环境，避免莫名其妙的问题", color("bold_yellow"))
+        shutil.rmtree(venv_path, ignore_errors=True)
 
     show_head_line("尝试初始化venv环境", color("bold_yellow"))
 
@@ -83,6 +84,7 @@ def parse_args():
     parser.add_argument("--venv_path", default=".venv")
     parser.add_argument("--requirements_path", default="requirements.txt")
     parser.add_argument("--dev", action="store_true")
+    parser.add_argument("--reset_venv", action="store_true")
     args = parser.parse_args()
 
     if args.dev:
@@ -94,4 +96,4 @@ def parse_args():
 
 if __name__ == "__main__":
     args = parse_args()
-    init_venv_and_requirements(args.venv_path, args.requirements_path, args.disable_douban, args.enable_proxy)
+    init_venv_and_requirements(args.venv_path, args.requirements_path, args.disable_douban, args.enable_proxy, args.reset_venv)
